requirements.txt
matplotlib==3.5.2
numpy==1.22.2
Pillow==9.1.1
requests==2.22.0
selenium==4.1.5
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
